Fix bug in calculation of interest.
Assume there was zero interest rate. Then the calculated interest was 11 * (the monthly difference in salaries) which was then added to your big comp salary. That is clearly wrong.

Instead of calculating the interest gain, calculate how much extra salary you are getting when you factor in interest from the startup salary and add it to that.